Adding the alert method for onSubmit and a button
