+++ my-app/src/index.js
@@ -3,13 +3,13 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 
 import Game from './app/screens/Game';
-import LogIn from './app/screens/LogIn';
+import RegisterFromContainer from './app/screens/LogIn';
 import './index.css';
 import { store } from './redux/store';
 
 const App = () => (
   <Provider store={store}>
-    <LogIn />
+    <RegisterFromContainer />
   </Provider>
 );
